refactor(cli/git): moves git logic to one file
